// src/StreamMaster.UI/lib/api.d.ts
export interface components {
	schemas: {
// ...
		M3UFileDto: components["schemas"]["BaseFileDto"] & {
			m3U8OutPutProfile: string;
			m3UName: components["schemas"]["M3UField"];
			m3UKey: components["schemas"]["M3UKey"];
			syncChannels: boolean;
			vodTags: string[];
			defaultStreamGroupName: string | null;
			/** Format: int32 */
			maxStreamCount: number;
			/** Format: int32 */
			streamCount: number;
			/** Format: int32 */
			startingChannelNumber: number;
			autoSetChannelNumbers: boolean;
		};
// ...
		CreateM3UFileRequest: {
			name?: string;
			/** Format: int32 */
			maxStreamCount?: number | null;
			m3U8OutPutProfile?: string | null;
			m3UKey?: components["schemas"]["M3UKey"] | null;
			m3UName?: components["schemas"]["M3UField"] | null;
			defaultStreamGroupName?: string | null;
			urlSource?: string | null;
			syncChannels?: boolean | null;
			/** Format: int32 */
			hoursToUpdate?: number | null;
			/** Format: int32 */
			startingChannelNumber?: number | null;
			autoSetChannelNumbers?: boolean | null;
			vodTags?: string[] | null;
		};
// ...
	};
	responses: never;
	parameters: never;
	requestBodies: never;
	headers: never;
	pathItems: never;
}